+++ wallets/provider-all/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [0.46.0](https://github.com/rango-exchange/rango-client/compare/provider-all@0.45.1...provider-all@0.46.0) (2025-05-26)
+
+
+
 ## [0.45.1](https://github.com/rango-exchange/rango-client/compare/provider-all@0.45.0...provider-all@0.45.1) (2025-05-04)